pyproject.toml
[tool.poetry]
name = "arrr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyodide-py = "^0.23.2"
arrr1 = "^1.0.1"
numpy = "^1.24.3"
matplotlib = "^3.7.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
